gds/requirements.txt
Flask==2.3.2
Flask-Sock==0.5.0
pyserial==3.5
Werkzeug==2.3.7
pytz